fix(avo-2211): extract translations, removed console.log, move config
